refactor: clean up legacy stores, remove hasContacts, and prune re-export hub
