+++ assistant/src/__tests__/send-endpoint-busy.test.ts
@@ -16,7 +16,7 @@ mock.module("../config/env.js", () => ({ isHttpAuthDisabled: () => true }));
 import type { ServerMessage } from "../daemon/ipc-protocol.js";
 import type { Session } from "../daemon/session.js";
 import { createCanonicalGuardianRequest } from "../memory/canonical-guardian-store.js";
-import { createBinding } from "../memory/channel-guardian-store.js";
+import { createBinding } from "../memory/guardian-bindings.js";
 import { getOrCreateConversation } from "../memory/conversation-key-store.js";
 
 const testDir = realpathSync(
