+++ assistant/src/__tests__/trusted-contact-multichannel.test.ts
@@ -96,7 +96,7 @@ mock.module("../runtime/approval-message-composer.js", () => ({
   composeApprovalMessageGenerative: async () => "mock generative message",
 }));
 
-import { createBinding } from "../memory/channel-guardian-store.js";
+import { createBinding } from "../memory/guardian-bindings.js";
 import { getDb, initializeDb, resetDb } from "../memory/db.js";
 import { findMember, upsertMember } from "../memory/ingress-member-store.js";
 import {
