+++ assistant/src/calls/relay-server.ts
@@ -21,7 +21,7 @@ import { contactChannelToMemberRecord } from "../contacts/member-record-shim.js"
 import { upsertMemberContactsFirst } from "../contacts/contacts-write.js";
 import { getAssistantName } from "../daemon/identity-helpers.js";
 import { getCanonicalGuardianRequest } from "../memory/canonical-guardian-store.js";
-import { listActiveBindingsByAssistant } from "../memory/channel-guardian-store.js";
+import { listActiveBindingsByAssistant } from "../memory/guardian-bindings.js";
 import * as conversationStore from "../memory/conversation-store.js";
 import { findActiveVoiceInvites } from "../memory/ingress-invite-store.js";
 import { revokeScopedApprovalGrantsForContext } from "../memory/scoped-approval-grants.js";
